refactor(ios): Move rendering out of `ViewController` This change moves the render methods out of `ViewController` and into static methods on a new `Renderer` struct with a view to ultimately moving device update management out of the view controller instance. This should allow renders when the full UI stack isn't present and decouple device updates from UI, making it easier to support multiple device configurations in the future.
